requirements-gpu.txt
tqdm
pillow
numpy
matplotlib
wandb
pytest
ruff

--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision